--- examples/input/af_output_analysis.yaml
@@ -1,9 +1,0 @@
-- structure_path: "/home/omkar/Omkar/omg_forks/IMP_Toolbox/examples/input/af_structure.cif"
-  data_path: "/home/omkar/Omkar/omg_forks/IMP_Toolbox/examples/input/af_data.json"
-  selection:
-    - id: "A"
-      model_region: [1,375]
-      af_region: [1,375]
-    - id: "B"
-      model_region: [1,127]
-      af_region: [1,127]
