--- examples/output/af_input/RNA_DNA_complex_8I54_set_0.json
@@ -1,126 +1,0 @@
-[
-    {
-        "name": "Lb2Cas12a_RNA_DNA_complex",
-        "modelSeeds": [
-            1
-        ],
-        "sequences": [
-            {
-                "proteinChain": {
-                    "sequence": "MYYESLTKQYPVSKTIRNELIPIGKTLDNIRQNNILESDVKRKQNYEHVKGILDEYHKQLINEALDNCTLPSLKIAAEIYLKNQKEVSDREDFNKTQDLLRKEVVEKLKAHENFTKIGKKDILDLLEKLPSISEDDYNALESFRNFYTYFTSYNKVRENLYSDKEKSSTVAYRLINENFPKFLDNVKSYRFVKTAGILADGLGEEEQDSLFIVETFNKTLTQDGIDTYNSQVGKINSSINLYNQKNQKANGFRKIPKMKMLYKQILSDREESFIDEFQSDEVLIDNVESYGSVLIESLKSSKVSAFFDALRESKGKNVYVKNDLAKTAMSNIVFENWRTFDDLLNQEYDLANENKKKDDKYFEKRQKELKKNKSYSLEHLCNLSEDSCNLIENYIHQISDDIENIIINNETFLRIVINEHDRSRKLAKNRKAVKAIKDFLDSIKVLERELKLINSSGQELEKDLIVYSAHEELLVELKQVDSLYNMTRNYLTKKPFSTEKVKLNFNRSTLLNGWDRNKETDNLGVLLLKDGKYYLGIMNTSANKAFVNPPVAKTEKVFKKVDYKLLPVPNQMLPKVFFAKSNIDFYNPSSEIYSNYKKGTHKKGNMFSLEDCHNLIDFFKESISKHEDWSKFGFKFSDTASYNDISEFYREVEKQGYKLTYTDIDETYINDLIERNELYLFQIYNKDFSMYSKGKLNLHTLYFMMLFDQRNIDDVVYKLNGEAEVFYRPASISEDELIIHKAGEEIKNKNPNRARTKETSTFSYDIVKDKRYSKDKFTLHIPITMNFGVDEVKRFNDAVNSAIRIDENVNVIGIDRGERNLLYVVVIDSKGNILEQISLNSIINKEYDIETDYHALLDEREGGRDKARKDWNTVENIRDLKAGYLSQVVNVVAKLVLKYNAIICLEDLNFGFKRGRQKVEKQVYQKFEKMLIDKLNYLVIDKSREQTSPKELGGALNALQLTSKFKSFKELGKQSGVIYYVPAYLTSKIDPTTGFANLFYMKCENVEKSKRFFDGFDFIRFNALENVFEFGFDYRSFTQRACGINSKWTVCTNGERIIKYRNPDKNNMFDEKVVVVTDEMKNLFEQYKIPYEDGRNVKDMIISNEEAEFYRRLYRLLQQTLQMRNSTSDGTRDYIISPVKNKREAYFNSELSDGSVPKDADANGAYNIARKGLWVLEQIRQKSEGEKINLAMTNAEWLEYAQTHLL",
-                    "glycans": [
-                        {
-                            "residues": "BMA",
-                            "position": 5
-                        }
-                    ],
-                    "modifications": [
-                        {
-                            "ptmType": "CCD_HY3",
-                            "ptmPosition": 11
-                        }
-                    ],
-                    "count": 1
-                }
-            },
-            {
-                "rnaSequence": {
-                    "sequence": "AAUUUCUACUAAUUGUAGAUGCCGCUACCCCGA",
-                    "modifications": [],
-                    "count": 1
-                }
-            },
-            {
-                "dnaSequence": {
-                    "sequence": "TGGTCGGGGTAGCGGCTAAAGCACT",
-                    "modifications": [],
-                    "count": 1
-                }
-            },
-            {
-                "dnaSequence": {
-                    "sequence": "AGTGCTTTA",
-                    "modifications": [
-                        {
-                            "modificationType": "CCD_6OG",
-                            "basePosition": 2
-                        },
-                        {
-                            "modificationType": "CCD_6MA",
-                            "basePosition": 1
-                        }
-                    ],
-                    "count": 1
-                }
-            },
-            {
-                "ion": {
-                    "ion": "MG",
-                    "count": 1
-                }
-            }
-        ]
-    },
-    {
-        "name": "Lb2Cas12a_RNA_DNA_complex",
-        "modelSeeds": [
-            2
-        ],
-        "sequences": [
-            {
-                "proteinChain": {
-                    "sequence": "MYYESLTKQYPVSKTIRNELIPIGKTLDNIRQNNILESDVKRKQNYEHVKGILDEYHKQLINEALDNCTLPSLKIAAEIYLKNQKEVSDREDFNKTQDLLRKEVVEKLKAHENFTKIGKKDILDLLEKLPSISEDDYNALESFRNFYTYFTSYNKVRENLYSDKEKSSTVAYRLINENFPKFLDNVKSYRFVKTAGILADGLGEEEQDSLFIVETFNKTLTQDGIDTYNSQVGKINSSINLYNQKNQKANGFRKIPKMKMLYKQILSDREESFIDEFQSDEVLIDNVESYGSVLIESLKSSKVSAFFDALRESKGKNVYVKNDLAKTAMSNIVFENWRTFDDLLNQEYDLANENKKKDDKYFEKRQKELKKNKSYSLEHLCNLSEDSCNLIENYIHQISDDIENIIINNETFLRIVINEHDRSRKLAKNRKAVKAIKDFLDSIKVLERELKLINSSGQELEKDLIVYSAHEELLVELKQVDSLYNMTRNYLTKKPFSTEKVKLNFNRSTLLNGWDRNKETDNLGVLLLKDGKYYLGIMNTSANKAFVNPPVAKTEKVFKKVDYKLLPVPNQMLPKVFFAKSNIDFYNPSSEIYSNYKKGTHKKGNMFSLEDCHNLIDFFKESISKHEDWSKFGFKFSDTASYNDISEFYREVEKQGYKLTYTDIDETYINDLIERNELYLFQIYNKDFSMYSKGKLNLHTLYFMMLFDQRNIDDVVYKLNGEAEVFYRPASISEDELIIHKAGEEIKNKNPNRARTKETSTFSYDIVKDKRYSKDKFTLHIPITMNFGVDEVKRFNDAVNSAIRIDENVNVIGIDRGERNLLYVVVIDSKGNILEQISLNSIINKEYDIETDYHALLDEREGGRDKARKDWNTVENIRDLKAGYLSQVVNVVAKLVLKYNAIICLEDLNFGFKRGRQKVEKQVYQKFEKMLIDKLNYLVIDKSREQTSPKELGGALNALQLTSKFKSFKELGKQSGVIYYVPAYLTSKIDPTTGFANLFYMKCENVEKSKRFFDGFDFIRFNALENVFEFGFDYRSFTQRACGINSKWTVCTNGERIIKYRNPDKNNMFDEKVVVVTDEMKNLFEQYKIPYEDGRNVKDMIISNEEAEFYRRLYRLLQQTLQMRNSTSDGTRDYIISPVKNKREAYFNSELSDGSVPKDADANGAYNIARKGLWVLEQIRQKSEGEKINLAMTNAEWLEYAQTHLL",
-                    "glycans": [
-                        {
-                            "residues": "BMA",
-                            "position": 5
-                        }
-                    ],
-                    "modifications": [
-                        {
-                            "ptmType": "CCD_HY3",
-                            "ptmPosition": 11
-                        }
-                    ],
-                    "count": 1
-                }
-            },
-            {
-                "rnaSequence": {
-                    "sequence": "AAUUUCUACUAAUUGUAGAUGCCGCUACCCCGA",
-                    "modifications": [],
-                    "count": 1
-                }
-            },
-            {
-                "dnaSequence": {
-                    "sequence": "TGGTCGGGGTAGCGGCTAAAGCACT",
-                    "modifications": [],
-                    "count": 1
-                }
-            },
-            {
-                "dnaSequence": {
-                    "sequence": "AGTGCTTTA",
-                    "modifications": [
-                        {
-                            "modificationType": "CCD_6OG",
-                            "basePosition": 2
-                        },
-                        {
-                            "modificationType": "CCD_6MA",
-                            "basePosition": 1
-                        }
-                    ],
-                    "count": 1
-                }
-            },
-            {
-                "ion": {
-                    "ion": "MG",
-                    "count": 1
-                }
-            }
-        ]
-    }
-]
