--- examples/output/af_output/af_structure.txt
@@ -1,4 +1,0 @@
-Rigid body 1
-A: 4-40,49-375
-B: 1-127
-
