--- examples/output/best_structures.csv
@@ -1,20 +1,0 @@
-uniprot_id,best_chain,best_structure,coverage,unp_start,unp_end,start,end,resolution
-Q02487,K,7a7d,0.622,136,679,1,544,26.0
-,B,5erp,0.509,236,680,1,445,2.7
-,A,5j5j,0.114,136,235,1,100,3.29
-Q14126,E,8qk3,1.0,1,1118,1,1118,3.2
-,A,8qjx,1.0,1,1118,1,1118,3.3
-,A,8qjy,1.0,1,1118,1,1118,3.5
-,A,5erd,0.505,50,602,1,553,2.9
-,B,7agf,0.217,149,386,6,243,3.1
-,B,7agg,0.217,149,386,6,243,3.3
-,B,6sit,0.217,149,386,1,238,4.5
-,E,6qnu,0.205,157,380,1,224,3.8
-,D,6qnt,0.19,157,380,1,208,3.5
-,A,2yqg,0.106,47,162,8,123,
-P14923,B,3ifq,0.742,124,676,1,553,2.8
-Q99959,A,3tt9,0.262,346,620,3,233,1.55
-P15924,A,5dzz,0.17,1960,2448,6,494,2.6
-,A,3r6n,0.157,178,627,1,450,2.95
-,A,1lm7,0.086,2209,2456,1,248,3.0
-,B,1lm5,0.075,2609,2822,1,214,1.8
