--- examples/output/rigid_bodies/dp1_1_1to2871_model_0.cxc
@@ -1,39 +1,0 @@
-open /home/omkar/Omkar/IMP_Toolbox/pre_processing/inputs/AF_predictions/dp1_1_1to2871/fold_dp1_1_1to2871_model_0.cif
-col #1/A:178-627 #899181
-col #1/A:659-1037 #40cdca
-col #1/A:1958-2205 #d4df9e
-col #1/A:2206-2448 #eb8829
-col #1/A:1741-1941 #fe2d4c
-col #1/A:2613-2809 #449331
-col #1/A:1587-1740 #cb7401
-col #1/A:1437-1586 #cfcae7
-col #1/A:1281-1387 #7d172b
-col #1/A:68-158 #99475b
-col #1/A:1212-1280 #bbb269
-col #1/A:1089-1153 #115174
-col #1/A:2501-2565 #b4215a
-col #1/A:1154-1211 #d6a0f7
-col #1/A:1038-1088 #e8f036
-col #1/A:1388-1436 #5f7590
-col #1/A:2462-2500 #7e0cbd
-col #1/A:33-54 #e9468d
-col #1/A:2580-2599 #dafd5f
-col #1/A:159-177 #99b320
-col #1/A:2835-2852 #0abb73
-col #1/A:2853-2870 #f42aba
-col #1/A:628-643 #312002
-col #1/A:644-658 #389720
-col #1/A:2820-2834 #f6dc1d
-col #1/A:0-13 #a04d2e
-col #1/A:2566-2579 #7eb5f7
-col #1/A:55-67 #38be62
-col #1/A:2449-2461 #001818
-col #1/A:2600-2612 #9d45b8
-col #1/A:22-32 #c143af
-col #1/A:1948-1957 #58bb45
-col #1/A:2810-2819 #438a49
-col #1/A:14-21 #eb2a10
-col #1/A:1942-1947 #c5ca64
-sel #1/A:1-584
-sel ~sel; hide sel r
-sel #1/A:71-89,105-153,179-208,210-584 ; sel ~sel ; col sel #FF0000
